Move resource definition functions into split file
